+++ apps/src/codebridge/Workspace/HeaderButtons.tsx
@@ -1,24 +1,15 @@
 import {sendCodebridgeAnalyticsEvent} from '@codebridge/utils/analyticsReporterHelper';
-import {DAPLink, WebUSB} from 'dapjs';
 import React, {useCallback} from 'react';
 
+import codebridgeI18n from '@cdo/apps/codebridge/locale';
 import {Button, buttonColors} from '@cdo/apps/componentLibrary/button';
 import {TooltipProps, WithTooltip} from '@cdo/apps/componentLibrary/tooltip';
 import {MAIN_PYTHON_FILE} from '@cdo/apps/lab2/constants';
 import {MultiFileSource} from '@cdo/apps/lab2/types';
 import VersionHistoryButton from '@cdo/apps/lab2/views/components/versionHistory/VersionHistoryButton';
 import {useDialogControl, DialogType} from '@cdo/apps/lab2/views/dialogs';
-import {
-  MICROBIT_VENDOR_ID,
-  MICROBIT_PRODUCT_ID,
-} from '@cdo/apps/maker/boards/microBit/MicroBitConstants';
-import {
-  detectMicroBitVersion,
-  flashHexString,
-  getModifiedMicroPythonHexFile,
-} from '@cdo/apps/maker/boards/microBit/utils';
+import {sendPythonCodeToMicroBit} from '@cdo/apps/maker/boards/microBit/utils';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
-import {PythonlabLevelProperties} from '@cdo/apps/pythonlab/types';
 import {useAppSelector} from '@cdo/apps/util/reduxHooks';
 import commonI18n from '@cdo/locale';
 
@@ -32,9 +23,7 @@ const WorkspaceHeaderButtons: React.FunctionComponent = () => {
 
   const appName = useAppSelector(state => state.lab.levelProperties?.appName);
   const enableMicroBit = useAppSelector(
-    state =>
-      (state.lab.levelProperties as PythonlabLevelProperties | undefined)
-        ?.enableMicroBit || false
+    state => state.lab.levelProperties?.enableMicroBit || false
   );
   const skipUrl = useAppSelector(state => state.lab.levelProperties?.skipUrl);
   const dialogControl = useDialogControl();
@@ -42,12 +31,6 @@ const WorkspaceHeaderButtons: React.FunctionComponent = () => {
     state => state.lab2Project.projectSource?.source
   ) as MultiFileSource | undefined;
   const files = source?.files || {};
-  let pythonCode = '';
-  for (const file of Object.values(files as object)) {
-    if (file.name === MAIN_PYTHON_FILE) {
-      pythonCode = file.contents;
-    }
-  }
 
   const feedbackTooltipProps: TooltipProps = {
     text: commonI18n.feedback(),
@@ -78,42 +61,20 @@ const WorkspaceHeaderButtons: React.FunctionComponent = () => {
   }, [appName, dialogControl, skipUrl]);
 
   const onClickFlash = async () => {
-    console.log('Flash file onto micro:bit');
+    let pythonCode = '';
+    for (const file of Object.values(files as object)) {
+      if (file.name === MAIN_PYTHON_FILE) {
+        pythonCode = file.contents;
+      }
+    }
     if (pythonCode.trim().length === 0) {
       console.log(
         'There is no python code from main.py to send to the micro:bit.'
       );
       return;
     }
-    const device = await navigator.usb.requestDevice({
-      filters: [{vendorId: MICROBIT_VENDOR_ID, productId: MICROBIT_PRODUCT_ID}],
-    });
-    const microBitVersion = detectMicroBitVersion(device);
-    if (!microBitVersion) {
-      throw new Error('micro:bit version not detected correctly.');
-    }
-
-    const transport = new WebUSB(device);
-    const target = new DAPLink(transport);
-    // For now, log flash progress in dev console.
-    target.on(DAPLink.EVENT_PROGRESS, progress => {
-      if (Math.floor(progress * 100) % 10 === 0) {
-        console.log('progress percent', Math.floor(progress * 100));
-      }
-      if (progress === 1) {
-        console.log('FLASH COMPLETE');
-      }
-    });
-    const hexStrWithFiles = await getModifiedMicroPythonHexFile(
-      pythonCode,
-      microBitVersion
-    );
-    try {
-      await flashHexString(hexStrWithFiles, target);
-    } catch (error) {
-      console.log(error);
-      return Promise.reject('Failed to send MicroPython program to micro:bit.');
-    }
+    console.log('Flash file onto micro:bit');
+    sendPythonCodeToMicroBit(pythonCode);
   };
 
   return (
@@ -125,7 +86,7 @@ const WorkspaceHeaderButtons: React.FunctionComponent = () => {
           size={'xs'}
           type={'tertiary'}
           color={buttonColors.white}
-          text={'Send to micro:bit'}
+          text={codebridgeI18n.sendToMicroBit()}
           className={darkModeStyles.tertiaryButton}
         />
       )}
